% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{checkRF}
\alias{checkRF}
\title{Checks if predicted value was correct in dataframe}
\usage{
checkRF(data_with_pred, target)
}
\arguments{
\item{data_with_pred}{Data with predicted value}

\item{target}{Target column name}
}
\value{
List of correct and incorrect preditions
}
\description{
Checks if predicted value was correct in dataframe
}
