% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{predictRF}
\alias{predictRF}
\title{Predict value using trees from random forest}
\usage{
predictRF(random_forest, data, isClassification, levelList)
}
\arguments{
\item{random_forest}{List of trees to predict target}

\item{data}{Data to do prediction on}

\item{isClassification}{If target is class or continous}

\item{levelList}{If classification list of target levels}
}
\value{
Data with predicted value binded as column
}
\description{
Predict value using trees from random forest
}
