% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{rpartRF}
\alias{rpartRF}
\title{Creates and retuns random forest}
\usage{
rpartRF(dataset, target, predictor_names, percent_predictors, num_trees,
  complex_param, min_split, min_bucket, max_depth, percent_obs)
}
\arguments{
\item{dataset}{data to create trees}

\item{target}{attribute to predict}

\item{predictor_names}{attribute names used to predict target}

\item{percent_predictors}{percent of user predictors per tree}

\item{num_trees}{number of trees in forest}

\item{complex_param}{cp param in rpart (complexity parameter. Any split that does not decrease the overall lack of fit by a factor of cp is not attempted. For instance, with anova splitting, this means that the overall R-squared must increase by cp at each step. The main role of this parameter is to save computing time by pruning off splits that are obviously not worthwhile. Essentially,the user informs the program that any split which does not improve the fit by cp will likely be pruned off by cross-validation, and that hence the program need not pursue it)}

\item{min_split}{minsplit param in rpart (the minimum number of observations that must exist in a node in order for a split to be attempted)}

\item{min_bucket}{minbucket param in rpart (the minimum number of observations in any terminal <leaf> node. If only one of minbucket or minsplit is specified, the code either sets minsplit to minbucket*3 or minbucket to minsplit/3, as appropriate)}

\item{max_depth}{maxdepth param in rpart (Set the maximum depth of any node of the final tree, with the root node counted as depth 0. Values greater than 30 rpart will give nonsense results on 32-bit machines)}

\item{percent_obs}{percent of observations taken into account in each tree}
}
\value{
random forest with num_trees decision trees as list
}
\description{
Creates and retuns random forest
}
